+++ src/containers/QuestionnairesPage/index.js
@@ -1,0 +1,7 @@
+import { connect } from "react-redux";
+import QuestionnairePage from "./components/Questionnaires";
+
+const mapStateToProps = (state, ownProps) => ({});
+const mapDispatchToProps = {};
+
+export default connect(mapStateToProps, mapDispatchToProps)(QuestionnairePage);
