+++ src/layouts/FullPageLayout.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import BaseLayout from "layouts/Base";
+import BaseLayout from "layouts/BaseLayout";
 
 const FullPageLayout = props =>
   <BaseLayout hasNav hasBreadcrumbs hasUtilityBtns>
