--- src/layouts/index.js
@@ -1,3 +1,0 @@
-export { default as FullPageLayout } from "layouts/FullPage";
-export { default as SidebarPageLayout } from "layouts/WithSideBar";
-export { default as BaseLayout } from "layouts/Base";
