--- src/pages/index.js
@@ -1,4 +1,0 @@
-export { default as QuestionnairePage } from "./Questionnaire";
-export { default as QuestionnaireCreatePage } from "./CreateQuestionnaire";
-export { default as QuestionnaireDesignPage } from "./DesignQuestionnaire";
-export { default as NotFound } from "./NotFound";
