--- src/reducers/questionnaire/index.js
@@ -1,9 +1,0 @@
-import { combineReducers } from "redux";
-
-import items from "reducers/questionnaire/items";
-
-export const questionnaire = combineReducers({
-  items
-});
-
-export default questionnaire;
