--- src/reducers/questionnaire/index.test.js
@@ -1,10 +1,0 @@
-import questionnaire from "reducers/questionnaire/";
-import { defaultState as items } from "reducers/questionnaire/items";
-
-describe("reducers/questionnaire", () => {
-  it("should return the expected default state from subreducers", function() {
-    expect(questionnaire(undefined, {})).toEqual({
-      items
-    });
-  });
-});
